Move functions related to board from views.py to views_board.py. Change urls from '/list' to '/board and from /listview -> /board/listview. Add class BoardView to urls.py.
