+++ templates/edit.html
@@ -46,8 +46,8 @@
                 </div>
             </div>
             <div class="bt_wrap">
-                <a href="view.html" class="on">수정</a>
-                <a href="view.html">취소</a>
+                <a href="/view.html" class="on">수정</a>
+                <a href="/view.html">취소</a>
             </div>
         </div>
     </div>
