+++ templates/profileupdate.html
@@ -49,7 +49,7 @@
                             <ul class="dropdown-menu" aria-labelledby="navbarDropdown">
                                 <li><a class="dropdown-item" href="#!">자주 묻는 질문</a></li>
                                 <li><a class="dropdown-item" href="#!">실시간 채팅 상담</a></li>
-                                <li><a class="dropdown-item" href="/list">고객 상담 게시판</a></li>
+                                <li><a class="dropdown-item" href="/board">고객 상담 게시판</a></li>
                             </ul>
                         </li>
                     </ul>
