+++ templates/view.html
@@ -48,8 +48,8 @@
                 </div>
             </div>
             <div class="bt_wrap">
-                <a href="/list" class="on">목록</a>
-                <a href="/edit">수정</a>
+                <a href="/board" class="on">목록</a>
+                <a href="/board/edit">수정</a>
             </div>
         </div>
     </div>
