+++ templates/write.html
@@ -36,8 +36,8 @@
                 </div>
             </div>
             <div class="bt_wrap">
-                <a href="/view" class="on">등록</a>
-                <a href="/list">취소</a>
+                <a href="/board" class="on">등록</a>
+                <a href="/board">취소</a>
             </div>
         </div>
     </div>
