+++ web/urls.py
@@ -1,6 +1,7 @@
 from django_request_mapping import UrlPattern
 
 from web.views import MyView
+from web.views_board import BoardView
 from web.views_rest import RestView
 from web.views_review import ReviewView
 
@@ -8,3 +9,4 @@ urlpatterns = UrlPattern();
 urlpatterns.register(MyView);
 urlpatterns.register(RestView);
 urlpatterns.register(ReviewView);
+urlpatterns.register(BoardView);
